Fix to style, fill screen.
Found the problem. It was the fixed width on the body tag. I don't remember why I put that there, but it seemed like I had a reason. In any case, that seems to have fixed the problem. I also made the menu a little taller. It was a pixel off on the computer I'm viewing it on, running vista. Let me know if it looks weird now on any other machines. I'll look at it a little later I'm sure.